pyproject.toml
[tool.poetry]
name = "Opencast"
version = "0.1.0"
description = "A client-server media player that transforms your Raspberry Pi into an awesome streaming device"
authors = [
"Luc Sinet <luc.sinet@gmail.com>"
]
license = "MIT"

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/Tastyep/Pi-OpenCast"
keywords = ['raspberry', 'opencast', 'chromecast']

[tool.poetry.dependencies]
python = "^3.7"
toml = "^0.9"
youtube-dl = ">= 2020.3.24"
pyyaml = "^5.3.1"
"hurry.filesize" = "^0.9"
transitions = "^0.8.1"
structlog = "^20.1.0"
colorama = "^0.4.3"
python-vlc = "^3.0.9113"
aiohttp = "^3.6.2"
aiohttp_cors = "^0.7.0"
janus = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-cov = "^2.4"
sphinx_rtd_theme = "^0.4.3"
sphinx = "^3.1.1"
mypy = "^0.780"
sphinx-autodoc-typehints = "^1.11.0"
sphinxcontrib-openapi = "^0.7.0"
black = "^19.10b0"
flake8 = "^3.8.3"
isort = "^5.4.2"
coverage = "^5.2.1"

[tool.poetry.scripts]
my-script = 'OpenCast:main'

[tool.isort]
profile = "black"
src_paths = ["OpenCast", "test"]
